## Part 3 - Animating the Poles
Now we want to animate the poles so they move from left to right when we start the game. The game made it looks like the bird is flying forward, but in fact, it's the poles that are moving, our bird can only move up and down.
In order to do this, we are going to add animations to our JavaScript file.

1. The first thing we will do is add querySelector on the elements, and initialize some variables including the bird's speed.