## Part4 - Re-generating Poles
In part 2, the poles disappear after moving to the left. The last thing we will do is to generate new poles by moving the pole elements to the right of the game area.The new poles should have a different height that is randomly generated, while the gap between the pair of poles are fixed.

So whenever the old poles disappear to the left side of the game area, generate new pair of poles. First, we need to determine whether the poles reached the left side of the game area. Since`polesCurrentPos` (used in the previous part) is the length between the rightmost game area and the pole, that means if `polesCurrentPos` is bigger than the game area's width, `containerWidth`, the poles are moving out of the game area.